# docs/04-接口与数据设计/openapi/internal-api.json
{
  "components": {
    "schemas": {
...
      "ApiResponseDepartmentResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DepartmentResponse"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseListDepartmentResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DepartmentResponse"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateDepartmentRequest": {
        "properties": {
          "deptCode": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "deptName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "managerName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "parentId": {
            "format": "int64",
            "type": "integer"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "deptCode",
          "deptName",
          "organizationId"
        ],
        "type": "object"
      },
...
      "DepartmentResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deptCode": {
            "type": "string"
          },
          "deptName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "level": {
            "format": "int32",
            "type": "integer"
          },
          "managerName": {
            "type": "string"
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "organizationName": {
            "type": "string"
          },
          "parentId": {
            "format": "int64",
            "type": "integer"
          },
          "path": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpdateDepartmentRequest": {
        "properties": {
          "deptName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "managerName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "parentId": {
            "format": "int64",
            "type": "integer"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "内部用户认证、RBAC、动态菜单、审批任务、系统治理及阶段1主数据接口。",
    "title": "SRM 内部管理端 API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/master-data/organizations/{id}/disable": {
      "post": {
        "operationId": "disable_7",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用组织",
        "tags": [
          "organization-controller"
        ]
      }
    },
...
    "/api/v1/system/departments": {
      "get": {
        "operationId": "listByOrganization",
        "parameters": [
          {
            "in": "query",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListDepartmentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "按组织查询部门列表",
        "tags": [
          "department-controller"
        ]
      },
      "post": {
        "operationId": "create_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDepartmentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "创建部门",
        "tags": [
          "department-controller"
        ]
      }
    },
    "/api/v1/system/departments/{id}": {
      "get": {
        "operationId": "get_5",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDepartmentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "获取部门详情",
        "tags": [
          "department-controller"
        ]
      },
      "put": {
        "operationId": "update_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDepartmentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "更新部门",
        "tags": [
          "department-controller"
        ]
      }
    },
    "/api/v1/system/departments/{id}/disable": {
      "post": {
        "operationId": "disable_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用部门",
        "tags": [
          "department-controller"
        ]
      }
    },
    "/api/v1/system/departments/{id}/enable": {
      "post": {
        "operationId": "enable_5",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "启用部门",
        "tags": [
          "department-controller"
        ]
      }
    },
...
    "/api/v1/system/dictionaries/{id}/disable": {
      "post": {
        "operationId": "disable_5",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable dictionary",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
...
    "/api/v1/system/document-templates/{id}/disable": {
      "post": {
        "operationId": "disable_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
...
    "/api/v1/system/number-rules/{id}/disable": {
      "post": {
        "operationId": "disable_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
...
    "/api/v1/system/positions/{id}/disable": {
      "post": {
        "operationId": "disable_2",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用岗位",
        "tags": [
          "position-controller"
        ]
      }
    },
...
    "/api/v1/system/roles/{id}/disable": {
      "post": {
        "operationId": "disable_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用角色",
        "tags": [
          "role-controller"
        ]
      }
    },
...
    "/api/v1/system/users/{id}/disable": {
      "post": {
        "operationId": "disable",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用用户",
        "tags": [
          "user-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "通过当前站点的同源网关访问",
      "url": "/"
    }
  ]
}